Vastly improved zoom values
Several things:
1) Zoom order now matches GIMP, Paint.NET, etc - high values are at the
top, low values at the bottom
2) Zoom values are no longer ratios of 1 (1:2, 2:1 etc).  They are a
hand-selected assortment of useful values, based somewhat off values in
GIMP, Paint.NET, and various parts of the LibreOffice suite.

+++ Modules/VBP_PublicVariables.bas
@@ -31,7 +31,7 @@ Public FilterBias As Long
 Public Type ZoomData
     ZoomCount As Byte
     ZoomArray() As Double
-    ZoomFactor() As Byte
+    ZoomFactor() As Double
 End Type
 
 Public Zoom As ZoomData
